code/requirements.txt
python-dotenv==1.0.1
qdrant-client[fastembed]==1.14.3
openai==1.76.0
